+++ demo/pages/correct-answer-toggle.js
@@ -1,9 +1,10 @@
-import Checkbox from 'material-ui/Checkbox';
+import Checkbox from '@material-ui/core/Checkbox';
 import React from 'react';
 import Toggle from '@pie-lib/correct-answer-toggle';
-import { withStyles } from 'material-ui/styles';
-import { FormGroup, FormControlLabel } from 'material-ui/Form';
-import Typography from 'material-ui/Typography';
+import { withStyles } from '@material-ui/core/styles';
+import FormGroup from '@material-ui/core/FormGroup';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import Typography from '@material-ui/core/Typography';
 
 import withRoot from '../src/withRoot';