+++ packages/config-ui/__mocks__/material-ui/Menu.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Menu = () => <div>Menu</div>;
-
-export default Menu; 
-
-export const MenuItem = () => <div>MenuItem</div>;