+++ packages/config-ui/__mocks__/material-ui/Radio.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const mock = jest.genMockFromModule('material-ui/Radio');
-
-module.exports = mock;