+++ packages/config-ui/__mocks__/material-ui/Switch.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const mock = jest.genMockFromModule('material-ui/Switch');
-
-module.exports = mock;